requirements.txt
torch
torchrl
tensordict
hydra-core
hydra-submitit-launcher
wandb
termcolor
imageio
mujoco
dm_control
myosuite
tqdm
gymnasium
matplotlib
pandas